feat: The FileIO class has been moved to the new file_handler
This class and its readLineByLinea method have been moved to the new library,
and the necessary changes have been made to ensure that the tests run
successfully.
