+++ src/common/data_provider/CMakeLists.txt
@@ -165,7 +165,7 @@ add_library(sysinfo STATIC
     ${SRC_FOLDER}/${RESOURCE_OBJ})
 
 target_include_directories(sysinfo PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/include/)
-target_link_libraries(sysinfo PUBLIC Filesystem)
+target_link_libraries(sysinfo PUBLIC Filesystem FileIO)
 
 if(CMAKE_SYSTEM_NAME STREQUAL "Windows")
   target_link_libraries(sysinfo PUBLIC psapi iphlpapi ws2_32)
