+++ src/common/data_provider/tests/sysInfoPackages/sysInfoPackagesPYPI_test.hpp
@@ -14,7 +14,7 @@
 
 #include "gtest/gtest.h"
 #include "MockFileSystem.hpp"
-#include "MockFileIO.hpp"
+#include "../../../file_handler/file_io/tests/mocks/MockFileIO.hpp"
 #include "packagesPYPI.hpp"
 
 class PYPITest : public ::testing::Test
