+++ src/common/file_handler/file_io/include/file_io.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <filesystem>
+#include <functional>
+#include <string>
+
+namespace file_io
+{
+    class FileIO
+    {
+        public:
+            void readLineByLine(const std::filesystem::path& filePath,
+            const std::function<bool(const std::string&)>& callback);
+    };
+}
