+++ src/common/file_handler/file_io/tests/mocks/MockFileIO.hpp
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "gtest/gtest.h"
+#include "gmock/gmock.h"
+#include <filesystem>
+
+class MockFileIO
+{
+    public:
+        MOCK_METHOD(void, readLineByLine, (const std::filesystem::path&, const std::function<bool(const std::string&)>&), ());
+};
